Grafica ok per schermi grandi. Chat da finire. Alberto gay

+++ app/src/main/java/com/example/simello/utils/UpdatePositionReceiver.java
@@ -1,16 +1,12 @@
 package com.example.simello.utils;
 
-import android.app.ProgressDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.util.Log;
-import android.widget.Toast;
 
-import com.example.simello.classiServer.InsertUserInput;
 import com.example.simello.classiServer.UpdatePositionInput;
-import com.example.simello.guanxy.GuanxyActivity;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.apache.http.HttpResponse;
@@ -37,7 +33,7 @@ public class UpdatePositionReceiver extends BroadcastReceiver
         GPSManager gpsManager = new GPSManager(context);
         if(gpsManager.canGetLocation())
         {
-            UpdatePositionInput updatePositionInput = new UpdatePositionInput(idUser, (long) gpsManager.getLatitude(), (long) gpsManager.getLongitude());
+            UpdatePositionInput updatePositionInput = new UpdatePositionInput(idUser,  gpsManager.getLatitude(),  gpsManager.getLongitude());
             connectAsyncTask connectAsyncTask = new connectAsyncTask("http://5.249.151.38:8080/guanxy/user/updatePosition");
             connectAsyncTask.execute(updatePositionInput);
         }
